insta-post/pyproject.toml
[project]
name = "insta-post"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]
